added ability to delete reviews

+++ server/api/review/[reviewid]/index.delete.ts
@@ -3,13 +3,22 @@ import { DeleteReviewReq } from "~/types"
 export default defineEventHandler(async (event) => {
     const reviewid = getRouterParam(event, 'reviewid')
     const body: DeleteReviewReq = await readBody(event)
+    const authHeader = getHeaders(event).authorization
     const config = useRuntimeConfig()
 	
     try {
-        await $fetch(`${config.public.backendBaseURL}/review/${reviewid}`, {
-            method: "DELETE",
-            body
-        })
+        if (authHeader) {
+            await $fetch(`${config.public.backendBaseURL}/review/${reviewid}`, {
+                method: "DELETE",
+                headers: {
+                    "Authorization": authHeader
+                },
+                body,
+            })
+        }
+        else {
+            throw createError({ message: "No auth token provided." })
+        }
     }
     catch (e: any) {
         throw createError(e)
